Move Scale link from Caffe function helpers to standard links.
